+++ CHANGELOG.md
@@ -11,12 +11,12 @@ All notable changes to this project will be documented in this file.
 - Progress: pre-format styles.
 
 ### Fixed
-- Compilation using `clang`
+- Compilation using `clang`.
 
 ## [1.0.0] - 2021-01-29
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/lem0nez/friendly-cli/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/lem0nez/friendly-cli/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/lem0nez/friendly-cli/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/lem0nez/friendly-cli/releases/tag/v1.0.0